requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
python-multipart
aiofiles
python-jose
passlib
bcrypt
